fix(oo-schemas): expiratioon timestamp as BigDecimal

+++ src/models.rs
@@ -107,7 +107,7 @@ pub struct OORequest {
     pub callback_recipient: String, 
     pub escalation_manager: String, 
     pub caller: String, 
-    pub expiration_timestamp: i64, 
+    pub expiration_timestamp: BigDecimal, 
     pub settled: Option<bool>,
     pub settlement_resolution: Option<bool>, 
     pub settle_caller: Option<String>,
